more stuff
added some comments, broke out utils module, added clamp function and move_toward actor method
